[Fixes] Since the remainder is a digit an int should suffice

+++ 0x08-palindrome_integer/0-is_palindrome.c
@@ -9,7 +9,7 @@
 unsigned long reversed(unsigned long n)
 {
 	unsigned long r = 0;
-	unsigned long remainder;
+	int remainder;
 
 	while (n)   /*Calculates the reversed number.*/
 	{
